chore: refactor to use vercel edge functions.
