+++ lib/generate-embeddings.ts
