--- nextjs-openai-doc-search.code-workspace
@@ -1,23 +1,0 @@
-{
-  "folders": [
-    {
-      "name": "project-root",
-      "path": "./"
-    },
-    {
-      "name": "next-app",
-      "path": "app"
-    },
-    {
-      "name": "supabase-functions",
-      "path": "supabase/functions"
-    }
-  ],
-  "settings": {
-    "files.exclude": {
-      "node_modules/": true,
-      "supabase/functions/": true,
-      "app/": true
-    }
-  }
-}
