+++ pages/docs/openai_embeddings.mdx
