--- supabase/functions/.env.example
@@ -1,2 +1,0 @@
-# Get your key at https://platform.openai.com/account/api-keys
-OPENAI_KEY=
