--- supabase/functions/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-  "deno.enable": true,
-  "deno.unstable": true
-}
